+++ src/const/ContactItem.js
@@ -1,0 +1,19 @@
+const contactSection = {
+  title: "Contact",
+  body: `
+  簡単ではありますが、お問い合わせのフォームをご用意しました。
+          ホームページの改善点や私について連絡がしたい方は、こちらからご連絡ください。
+  `,
+};
+const alerts = {
+  success: {
+    severity: "success",
+    msg: "メッセージありがとうございます。後ほどご連絡いたします。",
+  },
+  error: {
+    severity: "error",
+    msg: "エラーが発生しました。恐れいりますが、時間を空けてもう一度お問い合わせください。",
+  },
+};
+
+export { contactSection, alerts };
